feat: refactor validation utils
